[build] Add MacOS build support for aimet-torch, aimet-onnx (#6609)
* Add MacOS build support for aimet (torch, onnx)

- Added platform detection for MacOS
- Updated RPATH handling to use @loader_path on MacOS
- Added install_name_tool support for MacOS
- Skip CUDA architecture detection on MacOS
- Updated library extension detection
- Added ONNX Runtime platform tags for MacOS
- Fixed compiler detection to use CMAKE_CXX_COMPILER instead of hardcoded g++
- Fix unit tests for aimet-torch
- Fix unit tests for aimet-onnx
* Address PR review: clean up dependency version pins
- Move deepspeed to reqs_pip_test_torch.txt
- Use == wildcard syntax for bokeh and panel version pins

+++ TrainingExtensions/torch/src/AimetEncodingRescaler.cpp
@@ -39,7 +39,7 @@ public:
         {
             auto options = torch::TensorOptions().dtype(torch::kFloat32).device(input.device().type(),
                                                                                 input.device().index());
-            scaling_params = torch::empty({weight_scale.size()}, options);
+            scaling_params = torch::empty({static_cast<int64_t>(weight_scale.size())}, options);
         }
         else
         {
@@ -48,13 +48,13 @@ public:
             {
                 auto options = torch::TensorOptions().dtype(torch::kFloat32).device(input.device().type(),
                                                                                     input.device().index());
-                scaling_params = torch::empty({weight_scale.size()}, options);
+                scaling_params = torch::empty({static_cast<int64_t>(weight_scale.size())}, options);
             }
         }
         DlQuantization::ConvSpecArgs<float> encodingArgs = {.out_encoding_delta = static_cast<float>(out_enc.delta),
                                                               .out_encoding_offset = static_cast<float>(out_enc.offset),
                                                               .input_scale = input_scale,
-                                                              .bw = out_enc.bw,
+                                                              .bw = static_cast<uint8_t>(out_enc.bw),
                                                               .weight_scale = weight_scale.cast<std::vector<float>>(),
         };
